IGNITE-19675 [IEP-104] Implement WALIterator over ByteBuffer
