+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/wal/ByteBufferBackedDataInput.java
@@ -20,249 +20,155 @@ package org.apache.ignite.internal.processors.cache.persistence.wal;
 import java.io.DataInput;
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import org.apache.ignite.internal.processors.cache.persistence.wal.crc.FastCrc;
-import org.apache.ignite.internal.processors.cache.persistence.wal.crc.IgniteDataIntegrityViolationException;
 import org.jetbrains.annotations.NotNull;
 
 /**
  * ByteBuffer backed data input
  */
-public interface ByteBufferBackedDataInput extends DataInput {
+public abstract class ByteBufferBackedDataInput implements DataInput {
     /**
      * @return ByteBuffer hold by data input
      */
-    public ByteBuffer buffer();
+    public abstract ByteBuffer buffer();
 
     /**
      * ensure that requested count of byte is available in data input and will try to load data if not
      * @param requested Requested number of bytes.
      * @throws IOException If failed.
      */
-    public void ensure(int requested) throws IOException;
+    public abstract void ensure(int requested) throws IOException;
 
     /**
      * @return Position in the stream.
      */
-    public int position();
+    public int position() {
+        return buffer().position();
+    }
 
     /**
-     * @param skipCheck If CRC check should be skipped.
-     * @return autoclosable fileInput, after its closing crc32 will be calculated and compared with saved one
+     * {@inheritDoc}
      */
-    public Crc32CheckingDataInput startRead(boolean skipCheck);
+    @Override public int skipBytes(int n) throws IOException {
+        ensure(n);
+
+        buffer().position(buffer().position() + n);
+
+        return n;
+    }
 
     /**
-     * Checking of CRC32.
+     * {@inheritDoc}
      */
-    public class Crc32CheckingDataInput implements ByteBufferBackedDataInput, AutoCloseable {
-        /** */
-        private final FastCrc crc = new FastCrc();
+    @Override public void readFully(@NotNull byte[] b) throws IOException {
+        ensure(b.length);
 
-        /** Last calc position. */
-        private int lastCalcPosition;
+        buffer().get(b);
+    }
 
-        /** Skip crc check. */
-        private boolean skipCheck;
+    /**
+     * {@inheritDoc}
+     */
+    @Override public void readFully(@NotNull byte[] b, int off, int len) throws IOException {
+        ensure(len);
 
-        /** */
-        private ByteBufferBackedDataInput delegate;
+        buffer().get(b, off, len);
+    }
 
-        /**
-         */
-        public Crc32CheckingDataInput(ByteBufferBackedDataInput delegate, boolean skipCheck) {
-            this.delegate = delegate;
-            this.lastCalcPosition = position();
-            this.skipCheck = skipCheck;
-        }
+    /**
+     * {@inheritDoc}
+     */
+    @Override public boolean readBoolean() throws IOException {
+        return readByte() == 1;
+    }
 
-        /** {@inheritDoc} */
-        @Override public int position() {
-            return delegate.buffer().position();
-        }
+    /**
+     * {@inheritDoc}
+     */
+    @Override public byte readByte() throws IOException {
+        ensure(1);
 
-        /** {@inheritDoc} */
-        @Override public void ensure(int requested) throws IOException {
-            int available = buffer().remaining();
+        return buffer().get();
+    }
 
-            if (available >= requested)
-                return;
+    /**
+     * {@inheritDoc}
+     */
+    @Override public int readUnsignedByte() throws IOException {
+        return readByte() & 0xFF;
+    }
 
-            updateCrc();
+    /**
+     * {@inheritDoc}
+     */
+    @Override public short readShort() throws IOException {
+        ensure(2);
 
-            delegate.ensure(requested);
+        return buffer().getShort();
+    }
 
-            lastCalcPosition = 0;
-        }
+    /**
+     * {@inheritDoc}
+     */
+    @Override public int readUnsignedShort() throws IOException {
+        return readShort() & 0xFFFF;
+    }
 
-        /** {@inheritDoc} */
-        @Override public Crc32CheckingDataInput startRead(boolean skipCheck) {
-            return null;
-        }
+    /**
+     * {@inheritDoc}
+     */
+    @Override public char readChar() throws IOException {
+        ensure(2);
 
-        /** {@inheritDoc} */
-        @Override public void close() throws Exception {
-            updateCrc();
+        return buffer().getChar();
+    }
 
-            int val = crc.getValue();
+    /**
+     * {@inheritDoc}
+     */
+    @Override public int readInt() throws IOException {
+        ensure(4);
 
-            int writtenCrc = this.readInt();
+        return buffer().getInt();
+    }
 
-            if ((val ^ writtenCrc) != 0 && !skipCheck) {
-                // If it is a last message we will skip it (EOF will be thrown).
-                ensure(5);
+    /**
+     * {@inheritDoc}
+     */
+    @Override public long readLong() throws IOException {
+        ensure(8);
 
-                throw new IgniteDataIntegrityViolationException(
-                    "val: " + val + " writtenCrc: " + writtenCrc
-                );
-            }
-        }
+        return buffer().getLong();
+    }
 
-        /**
-         *
-         */
-        private void updateCrc() {
-            if (skipCheck)
-                return;
+    /**
+     * {@inheritDoc}
+     */
+    @Override public float readFloat() throws IOException {
+        ensure(4);
 
-            int oldPos = buffer().position();
+        return buffer().getFloat();
+    }
 
-            buffer().position(lastCalcPosition);
+    /**
+     * {@inheritDoc}
+     */
+    @Override public double readDouble() throws IOException {
+        ensure(8);
 
-            crc.update(delegate.buffer(), oldPos - lastCalcPosition);
+        return buffer().getDouble();
+    }
 
-            lastCalcPosition = oldPos;
-        }
+    /**
+     * {@inheritDoc}
+     */
+    @Override public String readLine() throws IOException {
+        throw new UnsupportedOperationException();
+    }
 
-        /** {@inheritDoc} */
-        @Override public int skipBytes(int n) throws IOException {
-            ensure(n);
-
-            int skipped = Math.min(buffer().remaining(), n);
-
-            buffer().position(buffer().position() + skipped);
-
-            return skipped;
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public void readFully(@NotNull byte[] b) throws IOException {
-            ensure(b.length);
-
-            buffer().get(b);
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public void readFully(@NotNull byte[] b, int off, int len) throws IOException {
-            ensure(len);
-
-            buffer().get(b, off, len);
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public boolean readBoolean() throws IOException {
-            return readByte() == 1;
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public byte readByte() throws IOException {
-            ensure(1);
-
-            return buffer().get();
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public int readUnsignedByte() throws IOException {
-            return readByte() & 0xFF;
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public short readShort() throws IOException {
-            ensure(2);
-
-            return buffer().getShort();
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public int readUnsignedShort() throws IOException {
-            return readShort() & 0xFFFF;
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public char readChar() throws IOException {
-            ensure(2);
-
-            return buffer().getChar();
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public int readInt() throws IOException {
-            ensure(4);
-
-            return buffer().getInt();
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public long readLong() throws IOException {
-            ensure(8);
-
-            return buffer().getLong();
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public float readFloat() throws IOException {
-            ensure(4);
-
-            return buffer().getFloat();
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public double readDouble() throws IOException {
-            ensure(8);
-
-            return buffer().getDouble();
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public String readLine() throws IOException {
-            throw new UnsupportedOperationException();
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override public String readUTF() throws IOException {
-            throw new UnsupportedOperationException();
-        }
-
-        /** {@inheritDoc} */
-        @Override public ByteBuffer buffer() {
-            return delegate.buffer();
-        }
+    /**
+     * {@inheritDoc}
+     */
+    @Override public String readUTF() throws IOException {
+        throw new UnsupportedOperationException();
     }
 }
